requirements_312.txt
# Requirements for Python 3.12
# Derived from requirements.txt - removes torch/transformers (too heavy, not used by app)
# and fixes versions incompatible with Python 3.12

ag-ui-protocol==0.1.10
aiohappyeyeballs==2.6.1
aiohttp==3.11.11
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.40.0
anyio==4.7.0
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.35.99
botocore==1.35.99
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
cohere==5.13.12
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
fastavro==1.9.7
fastmcp==2.3.4
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
flask-cors==5.0.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
genai-prices==0.0.52
gitdb==4.0.11
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
griffe==1.5.4
groq==0.13.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface_hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.13
langchain-community==0.3.13
langchain-core==0.3.28
langchain-google-genai==2.0.8
langchain-groq==0.2.3
langchain-openai==0.3.1
langchain-text-splitters==0.3.4
langgraph==0.2.60
langgraph-checkpoint==2.0.10
langgraph-prebuilt==0.1.7
langgraph-sdk==0.1.51
langsmith==0.2.3
logfire==2.12.1
logfire-api==2.12.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.2
matplotlib==3.10.0
mcp==1.9.4
mistralai==1.2.7
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
multitasking==0.0.11
mypy_extensions==1.0.0
narwhals==1.22.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.57.4
openapi-pydantic==0.5.1
opencv-python==4.10.0.84
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathvalidate==3.2.1
pdfminer.six==20231228
pdfplumber==0.11.4
Pillow==11.1.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.1.1
psycopg2==2.9.10
pydantic==2.10.4
pydantic-ai==1.48.0
pydantic-ai-slim==1.48.0
pydantic-evals==1.48.0
pydantic-graph==1.48.0
pydantic-settings==2.7.0
pydantic_core==2.27.2
PyMuPDF==1.25.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2024.2
pywin32==308
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.4
setuptools==75.8.0
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sse-starlette==2.2.1
starlette==0.41.3
streamlit==1.41.1
streamlit-autorefresh==1.0.1
sympy==1.13.3
tavily-python==0.5.0
tenacity==9.0.0
textblob==0.18.0
tiktoken==0.8.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
tweepy==4.15.0
typer==0.15.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
vaderSentiment==3.3.2
watchdog==6.0.0
Werkzeug==3.1.3
wrapt==1.17.2
xlsxwriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.51
zipp==3.21.0
zstandard==0.23.0
